requirements.txt
attrs==22.1.0
ExifRead==3.0.0
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
tomli==2.0.1
